+++ CHANGELOG.md
@@ -2,11 +2,31 @@
 
 ## [Unreleased]
 
+## [2.0.0] - 2024-06-25
+
+### Added
+
+- Java 21 support
+- new change log & release checkers
+- migrated tests to assertJ
+- added readFileToString() method to BuildUtils
+- added response body to get methods exception message in case status code was not 200
+- upgraded Google Java formatter from 1.17. to 1.22
+- upgraded Lombok from 1.18.24 to 1.18.32
+
+## [1.0.1] - 2023-12-18
+
+### Added
+
+- more common classes
+
 ## [1.0.0] - 2023-12-18
 
 ### Added
 
 - release version checker
 
-[unreleased]: https://github.com/mlieshoff/supercell-connectors/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/mlieshoff/supercell-connectors/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/mlieshoff/supercell-connectors/compare/v1.0.1...2.0.0
+[1.0.1]: https://github.com/mlieshoff/supercell-connectors/compare/v1.0.0...1.0.1
 [1.0.0]: https://github.com/mlieshoff/supercell-connectors/compare/v1.0.0...1.0.0